Data Cleaning and Similarity Measurements
To look at the dataset I used df.info() I first looked at the number of non-null items in the dataset and found that all the columns had 50000 values except Minor and Overall GPA. I also noticed that Date of Birth is of object data type when it should be datetime. To fix the incorrect data type, I used pd.to_datetime(df['Date of Birth'],errors='coerce') for the column ‘Date of Birth.’ After this I noticed that the maximum age of the students was 124 using the age.describe(). So I decided to dive deeper and checked the number of people with the max value and found that their date of birth is in 1900. I decided to move forward and drop these students since this data is old and will not aid in any research to be done with the data. After dropping and resetting this I found that OVERALL GPA has a lower count than date of birth and major specific GPA and major specific GPA has a max GPA of 32. First I addressed the unusual GPA and found the students with a GPA of over 4.0. I noticed there are 10 students with an odd GPA, and they have an academic standing of ERROR and they are all born before 2008, since their values are unusual, I dropped these students. Our dataset has dropped from 50000 to 49980 however, that is only 20 individuals which doesn’t negatively affect our data. I continued searching for any other data to be cleaned such as any misspelled words or duplicated data but this dataset didn’t have that noise. However, I did notice that the overall GPA did have non-null items but instead of removing these students, I looked at all the major-specific GPAs in comparison to the Overall GPA and found that these values do not differ significantly. Hence, I decided to simply make them the same by using df['Overall GPA'].fillna(df['Major-Specific GPA']).